requirements.txt
chromadb
fastapi
langchain
openai
pydantic
python-dotenv
unstructured